+++ SpeedyWeather/src/parameterizations/radiation/zenith.jl
@@ -183,6 +183,19 @@ function year_angle(::Type{T}, time::DateTime) where {T}
 end
 
 """$(TYPEDSIGNATURES)
+Angle of year [radians] of `orbit_time` for the solar zenith `S`, or of its initial time
+if the seasonal cycle is disabled, see `year_angle`."""
+year_angle(::Type{T}, S::AbstractZenith, orbit_time::DateTime) where {T} =
+    year_angle(T, S.seasonal_cycle ? orbit_time : S.initial_time[])
+
+"""$(TYPEDSIGNATURES)
+Solar declination angle δ [radians] at `orbit_time` for the solar zenith `S`, following its
+`solar_declination` and seasonal cycle settings. Can be used to synchronize other seasonal
+cycles with the solar zenith angle."""
+solar_declination(::Type{T}, S::AbstractZenith, orbit_time::DateTime) where {T} =
+    S.solar_declination(year_angle(T, S, orbit_time))
+
+"""$(TYPEDSIGNATURES)
 Fraction of day in `time` as angle in radians [-π...π], noon to noon, at longitude `λ`.
 Always calculated relative to the Earth calendar, so LENGTH_OF_DAY is used as constant.
 This is because `Dates` functions assume an Earth date. Length of the daily cycle is instead
@@ -212,8 +225,7 @@ function cos_zenith!(
         throw(DimensionMismatch(geometry.spectral_grid.grid, cos_zenith.grid))
 
     # g: angular fraction of year [0...2π] for Jan-01 to Dec-31
-    time_of_year = S.seasonal_cycle ? orbit_time : S.initial_time[]
-    g = year_angle(NF, time_of_year)
+    g = year_angle(NF, S, orbit_time)
 
     # time correction [radians] due to the equation of time (sunrise/set oscillation)
     tc = S.time_correction(g)
@@ -290,8 +302,7 @@ function cos_zenith!(
         throw(DimensionMismatch(geometry.spectral_grid.grid, cos_zenith.grid))
 
     # g: angular fraction of year [0...2π] for Jan-01 to Dec-31
-    time_of_year = S.seasonal_cycle ? orbit_time : S.initial_time[]
-    g = year_angle(NF, time_of_year)
+    g = year_angle(NF, S, orbit_time)
 
     # solar declination angle [radians] changing from tropic of cancer to capricorn
     # throughout the year measured by g [radians]